+++ lp2p/host.go
@@ -5,6 +5,7 @@ package lp2p
 import (
 	"context"
 	"fmt"
+	"strings"
 	"time"
 
 	"github.com/cometbft/cometbft/config"
@@ -16,6 +17,7 @@ import (
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/libp2p/go-libp2p/p2p/protocol/ping"
 	quic "github.com/libp2p/go-libp2p/p2p/transport/quic"
+	ma "github.com/multiformats/go-multiaddr"
 )
 
 // Host is a wrapper around the libp2p host.
@@ -24,10 +26,14 @@ import (
 type Host struct {
 	host.Host
 
+	config config.LibP2PConfig
+
 	// bootstrapPeers are initial peers specified in the address book
-	bootstrapPeers []BootstrapPeer
+	bootstrapPeers map[peer.ID]BootstrapPeer
 
 	logger log.Logger
+
+	peerFailureHandlers []func(id peer.ID, err error)
 }
 
 // BootstrapPeer initial peers to connect to
@@ -48,7 +54,11 @@ func NewHost(config *config.P2PConfig, nodeKey cmcrypto.PrivKey, logger log.Logg
 		return nil, fmt.Errorf("libp2p is disabled")
 	}
 
-	privateKey, err := privateKeyFromCosmosKey(nodeKey)
+	if err := config.LibP2PConfig.ValidateBasic(); err != nil {
+		return nil, fmt.Errorf("invalid libp2p config: %w", err)
+	}
+
+	privateKey, err := PrivateKeyFromCosmosKey(nodeKey)
 	if err != nil {
 		return nil, fmt.Errorf("failed to convert private key to libp2p: %w", err)
 	}
@@ -58,7 +68,7 @@ func NewHost(config *config.P2PConfig, nodeKey cmcrypto.PrivKey, logger log.Logg
 		return nil, fmt.Errorf("failed to convert %q to multiaddr: %w", config.ListenAddress, err)
 	}
 
-	bootstrapPeers, err := BootstrapPeersFromConfig(config)
+	bootstrapPeers, err := BootstrapPeersFromConfig(config.LibP2PConfig)
 	switch {
 	case err != nil:
 		return nil, fmt.Errorf("failed to decode bootstrap peers: %w", err)
@@ -97,6 +107,7 @@ func NewHost(config *config.P2PConfig, nodeKey cmcrypto.PrivKey, logger log.Logg
 
 	return &Host{
 		Host:           host,
+		config:         config.LibP2PConfig,
 		bootstrapPeers: bootstrapPeers,
 		logger:         logger,
 	}, nil
@@ -106,8 +117,13 @@ func (h *Host) AddrInfo() peer.AddrInfo {
 	return peer.AddrInfo{ID: h.ID(), Addrs: h.Addrs()}
 }
 
-func (h *Host) BootstrapPeers() []BootstrapPeer {
+func (h *Host) BootstrapPeers() map[peer.ID]BootstrapPeer {
 	return h.bootstrapPeers
+}
+
+func (h *Host) BootstrapPeer(id peer.ID) (BootstrapPeer, bool) {
+	bp, ok := h.bootstrapPeers[id]
+	return bp, ok
 }
 
 func (h *Host) Logger() log.Logger {
@@ -122,31 +138,68 @@ func (h *Host) Ping(ctx context.Context, addrInfo peer.AddrInfo) (time.Duration,
 	return res.RTT, res.Error
 }
 
-func BootstrapPeersFromConfig(config *config.P2PConfig) ([]BootstrapPeer, error) {
-	peers := make([]BootstrapPeer, 0, len(config.LibP2PConfig.BootstrapPeers))
+func (h *Host) AddPeerFailureHandler(handler func(id peer.ID, err error)) {
+	h.peerFailureHandlers = append(h.peerFailureHandlers, handler)
+}
 
-	// dedup
-	cache := make(map[peer.ID]struct{})
+// EmitPeerFailure emits a peer failure event to all registered handlers.
+// This semantic is over host.eventBus for simplicity.
+func (h *Host) EmitPeerFailure(id peer.ID, err error) {
+	for _, handler := range h.peerFailureHandlers {
+		go handler(id, err)
+	}
+}
 
-	for _, bp := range config.LibP2PConfig.BootstrapPeers {
+func (h *Host) multiAddrStrByID(id peer.ID) string {
+	return multiAddrStr(h.Peerstore().Addrs(id))
+}
+
+func BootstrapPeersFromConfig(config config.LibP2PConfig) (map[peer.ID]BootstrapPeer, error) {
+	peers := make(map[peer.ID]BootstrapPeer, len(config.BootstrapPeers))
+
+	for _, bp := range config.BootstrapPeers {
 		addr, err := AddrInfoFromHostAndID(bp.Host, bp.ID)
 		if err != nil {
 			return nil, fmt.Errorf("[%s, %s]: %w", bp.Host, bp.ID, err)
 		}
 
-		if _, ok := cache[addr.ID]; ok {
+		if _, ok := peers[addr.ID]; ok {
 			continue
 		}
 
-		peers = append(peers, BootstrapPeer{
+		peers[addr.ID] = BootstrapPeer{
 			AddrInfo:      addr,
 			Private:       bp.Private,
 			Persistent:    bp.Persistent,
 			Unconditional: bp.Unconditional,
-		})
-
-		cache[addr.ID] = struct{}{}
+		}
 	}
 
 	return peers, nil
+}
+
+// IsDNSAddr checks if the given multiaddr is a DNS address.
+func IsDNSAddr(addr ma.Multiaddr) bool {
+	for _, a := range addr {
+		code := a.Protocol().Code
+
+		if code == ma.P_DNS || code == ma.P_DNS4 || code == ma.P_DNS6 || code == ma.P_DNSADDR {
+			return true
+		}
+	}
+
+	return false
+}
+
+func multiAddrStr(addrs []ma.Multiaddr) string {
+	if len(addrs) == 0 {
+		return "<empty>"
+	}
+
+	parts := make([]string, len(addrs))
+	for i, addr := range addrs {
+		parts[i] = addr.String()
+	}
+
+	return strings.Join(parts, ", ")
 }